+++ server/restAPI/types/backendData.go
@@ -3,6 +3,6 @@ package types
 type BackendData struct {
 	Name    string `json:"backend_name"`
 	ID      string `json:"id"`
-	Pointer uint32 `json:"pointer"`
+	Pointer uint64 `json:"pointer"`
 	Plugin  string `json:"plugin"`
 }
